+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Bug #62: Reorder use statements in `BaseInput` class for better organization (@terabytesoftw)
 - Bug #63: Use `HasName` trait in `BaseInput` class to manage the `name` attribute (@terabytesoftw)
 - Enh #64: Add support for automatic `aria-describedby` attribute in `BaseInput` class (@terabytesoftw)
+- Enh #65: Add stubs for `TagInline` and `TagInput` with token values for template rendering tests (@terabytesoftw)
 
 ## 0.5.2 January 28, 2026